tokenize "runners" for bash
