+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@moondreamsdev/dreamer-ui",
-  "version": "1.7.10-test.28",
+  "version": "1.7.10-test.29",
   "publishConfig": {
     "access": "public"
   },
